Conflicts:
	driver/event_table.c
	driver/flags_table.c
	driver/ppm_events.c
	driver/ppm_events_public.h
	driver/ppm_fillers.c
	userspace/libscap/CMakeLists.txt
	userspace/libscap/event_table.c
	userspace/libscap/flags_table.c
	userspace/libsinsp/event.cpp

+++ userspace/libscap/scap_fds.c
@@ -195,7 +195,7 @@ uint32_t scap_fd_info_len(scap_fdinfo *fdi)
 		res += 
 			sizeof(uint64_t) + // unix source 
 			sizeof(uint64_t) +  // unix destination
-			strnlen(fdi->info.unix_socket_info.fname, SCAP_MAX_PATH_SIZE) + 2;
+			(uint32_t)strnlen(fdi->info.unix_socket_info.fname, SCAP_MAX_PATH_SIZE) + 2;
 		break;
 	case SCAP_FD_FIFO:
 	case SCAP_FD_FILE:
@@ -206,7 +206,7 @@ uint32_t scap_fd_info_len(scap_fdinfo *fdi)
 	case SCAP_FD_EVENTPOLL:
 	case SCAP_FD_INOTIFY:
 	case SCAP_FD_TIMERFD:
-		res += strnlen(fdi->info.fname, SCAP_MAX_PATH_SIZE) + 2;    // 2 is the lenght field before the string
+		res += (uint32_t)strnlen(fdi->info.fname, SCAP_MAX_PATH_SIZE) + 2;    // 2 is the lenght field before the string
 		break;
 	default:
 		ASSERT(false);
@@ -279,7 +279,7 @@ int32_t scap_fd_write_to_disk(scap_t *handle, scap_fdinfo *fdi, gzFile f)
 			snprintf(handle->m_lasterr, SCAP_LASTERR_SIZE, "error writing to file (fi4)");
 			return SCAP_FAILURE;
 		}
-		stlen = strnlen(fdi->info.unix_socket_info.fname, SCAP_MAX_PATH_SIZE);
+		stlen = (uint16_t)strnlen(fdi->info.unix_socket_info.fname, SCAP_MAX_PATH_SIZE);
 		if(gzwrite(f, &stlen, sizeof(uint16_t)) != sizeof(uint16_t) ||
 		        (stlen > 0 && gzwrite(f, fdi->info.unix_socket_info.fname, stlen) != stlen))
 		{
@@ -296,7 +296,7 @@ int32_t scap_fd_write_to_disk(scap_t *handle, scap_fdinfo *fdi, gzFile f)
 	case SCAP_FD_EVENTPOLL:
 	case SCAP_FD_INOTIFY:
 	case SCAP_FD_TIMERFD:
-		stlen = strnlen(fdi->info.fname, SCAP_MAX_PATH_SIZE);
+		stlen = (uint16_t)strnlen(fdi->info.fname, SCAP_MAX_PATH_SIZE);
 		if(gzwrite(f, &stlen,  sizeof(uint16_t)) != sizeof(uint16_t) ||
 		        (stlen > 0 && gzwrite(f, fdi->info.fname, stlen) != stlen))
 		{
@@ -315,7 +315,7 @@ int32_t scap_fd_write_to_disk(scap_t *handle, scap_fdinfo *fdi, gzFile f)
 uint32_t scap_fd_read_prop_from_disk(scap_t *handle, OUT void *target, size_t expected_size, OUT size_t *nbytes, gzFile f)
 {
 	size_t readsize;
-	readsize = gzread(f, target, expected_size);
+	readsize = gzread(f, target, (unsigned int)expected_size);
 	CHECK_READ_SIZE(readsize, expected_size);
 	(*nbytes) += readsize;
 	return SCAP_SUCCESS;
